--- a/bench/index.js
+++ b/bench/index.js
@@ -2,7 +2,7 @@ var fs = require('fs')
 var path = require('path')
 var geojsonVt = require('geojson-vt')
 var Pbf = require('pbf')
-var VectorTile = require('vector-tile').VectorTile
+var VectorTile = require('@mapbox/vector-tile').VectorTile
 var Benchmark = require('benchmark')
 var serialize = require('../')