Replace generated code that requires lots of temporary allocations with hand-written streaming writes. This improves performance significantly for sources with many features by reducing GC pressure. (Although surprisingly it seems to slow very simple sources down slightly. I'm not sure why.)

Before:

```
raw x 92,357 ops/sec ±0.88% (93 runs sampled)
simple x 97,692 ops/sec ±1.04% (94 runs sampled)
points x 114 ops/sec ±0.77% (76 runs sampled)
lots of properties x 2,861 ops/sec ±1.32% (91 runs sampled)
```

After:

```
raw x 82,764 ops/sec ±1.38% (90 runs sampled)
simple x 87,773 ops/sec ±1.05% (89 runs sampled)
points x 176 ops/sec ±1.11% (84 runs sampled)
lots of properties x 2,849 ops/sec ±1.08% (94 runs sampled)
```

Refs https://github.com/mapbox/mapbox-gl-js/issues/5208.