@jfirebaugh 2017-09-06
Updated and rebased.

```
raw x 91,286 ops/sec ±1.68% (88 runs sampled)
simple x 96,556 ops/sec ±1.30% (85 runs sampled)
points x 238 ops/sec ±1.88% (85 runs sampled)
lots of properties x 2,701 ops/sec ±1.41% (92 runs sampled)
```